feat(index.js): 게임 지우는 로직 완료
게임이 끝난 상태이고 올바른 비밀번호를 입력하면 게임을 지우는 로직 완료

+++ src/main/java/chess/dao/GameDao.java
@@ -80,7 +80,7 @@ public class GameDao {
                 new GameDto(resultSet.getInt("id"), resultSet.getString("name")));
     }
 
-    public String findPassword(int gameId) {
+    public String findPassword(long gameId) {
         String sql = "SELECT password FROM game WHERE id = :gameId";
         SqlParameterSource sqlParameterSource = new MapSqlParameterSource("gameId", gameId);
         return namedParameterJdbcTemplate.queryForObject(sql, sqlParameterSource, String.class);
